Extract connection lifecycle from DaemonStatus into ConnectionManager
DaemonStatus was a mixed-concern object handling both observable state
and connection lifecycle (connect, disconnect, reconfigure, auto-wake,
HTTPTransport ownership). Extract the lifecycle into ConnectionManager
so DaemonStatus is a pure state observer.

ConnectionManager (272 lines):
- Owns HTTPTransport (health checks + host tool execution)
- connect/disconnect/reconfigure with auto-wake (macOS)
- Bearer token updates
- Exposes callbacks for DaemonStatus to wire state updates

DaemonStatus (362 lines, down from 552):
- @Published state only (isConnected, daemonVersion, model info, etc.)
- Wires ConnectionManager callbacks to update published properties
- Forwards connect/disconnect/subscribe/send to owned components
- Message pre-processor for SSE-driven state updates

No consumer API changes — DaemonStatus forwarding accessors maintain
backward compatibility for httpTransport, config, isConnecting,
wakeHandler, recoveryPlatform, recoveryDeviceId.
