+++ clients/shared/Network/DaemonStatus.swift
@@ -3,12 +3,6 @@ import os
 
 private let log = Logger(subsystem: Bundle.main.bundleIdentifier ?? "com.vellum.vellum-assistant", category: "DaemonStatus")
 
-/// Shared signpost log for network instrumentation (Points of Interest lane in Instruments).
-private let networkLog = OSLog(
-    subsystem: Bundle.main.bundleIdentifier ?? "com.vellum.vellum-assistant",
-    category: .pointsOfInterest
-)
-
 /// Protocol for daemon status and connection management, enabling dependency injection and testing.
 @MainActor
 public protocol DaemonStatusProtocol: AnyObject {
@@ -24,25 +18,18 @@ public protocol DaemonStatusProtocol: AnyObject {
 /// Observable status of the assistant daemon connection. Publishes connection state,
 /// daemon version, model info, and other status properties derived from SSE events.
 ///
-/// Replaces `DaemonClient` as the single observable object for daemon state.
-/// Owns `EventStreamClient` (SSE + subscribe) and `HTTPTransport` (health checks).
-/// SwiftUI views observe this for connection status and daemon metadata.
+/// Pure state object — connection lifecycle is managed by `ConnectionManager`,
+/// SSE and message broadcast by `EventStreamClient`. This class wires the three
+/// together and reacts to events by updating `@Published` properties.
 @MainActor
 public final class DaemonStatus: ObservableObject, DaemonStatusProtocol {
 
     // MARK: - Published State
 
     @Published public var isConnected: Bool = false
-    public var isConnecting: Bool = false
 
     /// The runtime HTTP server port, populated via `daemon_status` on connect.
     @Published public var httpPort: Int?
-
-    /// Platform identifier for automatic 401 re-bootstrap (e.g. "macos", "ios").
-    public var recoveryPlatform: String?
-
-    /// Device identifier for automatic 401 re-bootstrap.
-    public var recoveryDeviceId: String?
 
     /// Returns a closure that resolves the current HTTP port at call time.
     public var httpPortResolver: () -> Int? {
@@ -99,38 +86,105 @@ public final class DaemonStatus: ObservableObject, DaemonStatusProtocol {
         }
     }
 
-    var isAuthenticated = false
-
-    // MARK: - Auto-Wake
-
-    /// Optional closure invoked when a connection attempt fails because the daemon process
-    /// is not alive. The macOS app sets this to call `vellumCli.wake(name:)`.
-    public var wakeHandler: (@MainActor @Sendable () async throws -> Void)?
-
-    #if os(macOS)
-    var lastAutoWakeAttempt: Date?
-    var autoWakeTask: Task<Void, Never>?
-    #endif
-
-    // MARK: - Connection
-
-    /// HTTP transport for health checks and host tool execution.
-    public var httpTransport: HTTPTransport?
-
-    public private(set) var config: DaemonConfig
+    // MARK: - Components
 
     /// Event stream client for SSE and message broadcast.
     public let eventStreamClient: EventStreamClient
 
+    /// Connection lifecycle manager (health checks, auto-wake, transport).
+    public let connectionManager: ConnectionManager
+
+    // MARK: - Forwarding accessors (backward compat)
+
+    /// Whether a connection attempt is in flight.
+    public var isConnecting: Bool {
+        get { connectionManager.isConnecting }
+        set { connectionManager.isConnecting = newValue }
+    }
+
+    /// HTTP transport for health checks and host tool execution.
+    public var httpTransport: HTTPTransport? {
+        connectionManager.httpTransport
+    }
+
+    /// Current daemon connection configuration.
+    public var config: DaemonConfig {
+        connectionManager.config
+    }
+
+    /// Platform identifier for automatic 401 re-bootstrap.
+    public var recoveryPlatform: String? {
+        get { connectionManager.recoveryPlatform }
+        set { connectionManager.recoveryPlatform = newValue }
+    }
+
+    /// Device identifier for automatic 401 re-bootstrap.
+    public var recoveryDeviceId: String? {
+        get { connectionManager.recoveryDeviceId }
+        set { connectionManager.recoveryDeviceId = newValue }
+    }
+
+    /// Optional closure invoked when the daemon process is not alive.
+    public var wakeHandler: (@MainActor @Sendable () async throws -> Void)? {
+        get { connectionManager.wakeHandler }
+        set { connectionManager.wakeHandler = newValue }
+    }
+
     // MARK: - Init
 
     public init(config: DaemonConfig = .default) {
-        self.config = config
-        self.eventStreamClient = EventStreamClient()
+        let esc = EventStreamClient()
+        self.eventStreamClient = esc
+        self.connectionManager = ConnectionManager(config: config, eventStreamClient: esc)
 
         // Wire the pre-processor so state is updated before subscribers see messages
-        eventStreamClient.messagePreProcessor = { [weak self] message in
+        esc.messagePreProcessor = { [weak self] message in
             self?.handleServerMessage(message)
+        }
+
+        // Wire conversation ID resolution to subscribers.
+        esc.onConversationIdResolved = { [weak esc] localId, serverId in
+            esc?.broadcastMessage(.conversationIdResolved(localId: localId, serverId: serverId))
+        }
+
+        // Persist refreshed bearer tokens so the client survives app restarts.
+        esc.onTokenRefreshed = { newToken in
+            #if os(iOS)
+            let _ = APIKeyManager.shared.setAPIKey(newToken, provider: "runtime-bearer-token")
+            #elseif os(macOS)
+            // macOS re-reads from disk on each request; no persistence needed here.
+            #endif
+        }
+
+        // Wire ConnectionManager callbacks to update DaemonStatus state.
+        connectionManager.onConnectionStateChanged = { [weak self] connected in
+            guard let self else { return }
+            self.isConnected = connected
+            if connected {
+                NotificationCenter.default.post(name: .daemonDidReconnect, object: self)
+            }
+        }
+
+        connectionManager.onDaemonVersionChanged = { [weak self] newVersion in
+            guard let self else { return }
+            self.daemonVersion = newVersion
+            self.checkVersionCompatibility(daemonVersion: newVersion)
+            if self.isUpdateInProgress {
+                if newVersion == self.updateTargetVersion {
+                    log.info("Health check confirmed update completed — now running \(newVersion, privacy: .public)")
+                } else {
+                    log.warning("Health check detected version \(newVersion, privacy: .public) after update — expected \(self.updateTargetVersion ?? "?", privacy: .public), may have rolled back")
+                }
+                self.isUpdateInProgress = false
+                self.updateTargetVersion = nil
+                self.updateExpiresAt = nil
+                self.connectionManager.httpTransport?.setUpdateInProgress(false)
+                self.eventStreamClient.resetSSEReconnectDelay()
+            }
+        }
+
+        connectionManager.onAuthError = { [weak self] message in
+            self?.eventStreamClient.broadcastMessage(message)
         }
     }
 
@@ -169,19 +223,24 @@ public final class DaemonStatus: ObservableObject, DaemonStatusProtocol {
         eventStreamClient.stopSSE()
     }
 
-    // MARK: - Reconfigure
+    // MARK: - Connection (forwarding)
 
-    /// Reconfigure the daemon status transport in place without replacing
-    /// the object identity. Preserves subscriber references held by long-lived
-    /// objects across assistant switches.
+    public func connect() async throws {
+        try await connectionManager.connect()
+    }
+
+    public func disconnect() {
+        connectionManager.disconnect()
+        isConnected = false
+        httpPort = nil
+        latestMemoryStatus = nil
+    }
+
+    /// Reconfigure the transport in place without replacing object identity.
     public func reconfigure(config newConfig: DaemonConfig) {
-        #if os(macOS)
-        autoWakeTask?.cancel()
-        autoWakeTask = nil
-        #endif
-        disconnect()
-        self.config = newConfig
-        isAuthenticated = false
+        connectionManager.reconfigure(config: newConfig)
+        // Reset connection-specific published state
+        isConnected = false
         httpPort = nil
         daemonVersion = nil
         versionMismatch = false
@@ -191,248 +250,11 @@ public final class DaemonStatus: ObservableObject, DaemonStatusProtocol {
         keyFingerprint = nil
         latestMemoryStatus = nil
         currentModel = nil
-        #if os(macOS)
-        lastAutoWakeAttempt = nil
-        #endif
     }
-
-    // MARK: - Connect
-
-    public func connect() async throws {
-        try await connectImpl(cancelAutoWake: true)
-    }
-
-    private func connectImpl(cancelAutoWake: Bool) async throws {
-        disconnectInternal(triggerReconnect: false, cancelAutoWake: cancelAutoWake)
-
-        isConnecting = true
-
-        guard case .http(let baseURL, let bearerToken, let conversationKey) = config.transport else {
-            isConnecting = false
-            log.info("connect: non-HTTP transport, skipping")
-            return
-        }
-
-        log.info("connect: establishing HTTP transport to \(baseURL, privacy: .public)")
-
-        let transport = HTTPTransport(
-            baseURL: baseURL,
-            bearerToken: bearerToken,
-            conversationKey: conversationKey,
-            transportMetadata: config.transportMetadata
-        )
-
-        // Bridge HTTP transport connection state (health-check driven) to DaemonStatus.
-        transport.onConnectionStateChanged = { [weak self] connected in
-            guard let self else { return }
-            self.isConnected = connected
-            self.isConnecting = false
-            if connected {
-                NotificationCenter.default.post(name: .daemonDidReconnect, object: self)
-            }
-            #if os(macOS)
-            if !connected {
-                self.autoWakeIfDaemonDied()
-            }
-            #endif
-        }
-
-        // Wire conversation ID resolution from EventStreamClient to subscribers.
-        // Only clean up the SSE remapping entry when a subscriber actually handles
-        // the resolution (on macOS, ConversationManager updates the VM's conversationId).
-        // On iOS, no handler exists — the mapping and synthetic ID must stay so
-        // parseSSEData can continue remapping for the unchanged synthetic ID.
-        eventStreamClient.onConversationIdResolved = { [weak self] localId, serverId in
-            guard let self else { return }
-            self.eventStreamClient.broadcastMessage(.conversationIdResolved(localId: localId, serverId: serverId))
-        }
-
-        // Persist refreshed bearer tokens so the client survives app restarts.
-        eventStreamClient.onTokenRefreshed = { newToken in
-            #if os(iOS)
-            let _ = APIKeyManager.shared.setAPIKey(newToken, provider: "runtime-bearer-token")
-            #elseif os(macOS)
-            // macOS re-reads from disk on each request; no persistence needed here.
-            #endif
-        }
-
-        // Sync daemon version from health checks and confirm update completion.
-        transport.onDaemonVersionChanged = { [weak self] newVersion in
-            guard let self else { return }
-            self.daemonVersion = newVersion
-            self.checkVersionCompatibility(daemonVersion: newVersion)
-            if self.isUpdateInProgress {
-                if newVersion == self.updateTargetVersion {
-                    log.info("Health check confirmed update completed — now running \(newVersion, privacy: .public)")
-                } else {
-                    log.warning("Health check detected version \(newVersion, privacy: .public) after update — expected \(self.updateTargetVersion ?? "?", privacy: .public), may have rolled back")
-                }
-                self.isUpdateInProgress = false
-                self.updateTargetVersion = nil
-                self.updateExpiresAt = nil
-                self.httpTransport?.setUpdateInProgress(false)
-                self.eventStreamClient.resetSSEReconnectDelay()
-            }
-        }
-
-        // Broadcast auth errors from health check 401 handling to subscribers.
-        transport.onAuthError = { [weak self] message in
-            guard let self else { return }
-            self.eventStreamClient.broadcastMessage(message)
-        }
-
-        self.httpTransport = transport
-
-        // Propagate update-in-progress state to new transport
-        if isUpdateInProgress {
-            transport.isUpdateInProgress = true
-        }
-
-        // Register the conversation key for host tool filtering
-        if !conversationKey.isEmpty {
-            eventStreamClient.registerConversationId(conversationKey)
-        }
-
-        do {
-            try await transport.connect()
-            isAuthenticated = true
-            isConnecting = false
-            log.info("connect: transport connected successfully to \(baseURL, privacy: .public)")
-
-            // Auto-start SSE now that health check passed
-            eventStreamClient.startSSE()
-        } catch {
-            #if os(macOS)
-            guard !Task.isCancelled else {
-                isConnecting = false
-                httpTransport = nil
-                log.info("connect: task cancelled — skipping auto-wake")
-                throw error
-            }
-
-            if let wakeHandler, config.transportMetadata.routeMode == .runtimeFlat {
-                let reachable = await HealthCheckClient.isReachable(instanceDir: config.instanceDir)
-                if !reachable {
-                    log.info("connect: gateway unreachable — attempting auto-wake before retry")
-                    do {
-                        try await wakeHandler()
-                        log.info("connect: auto-wake succeeded, retrying connection to \(baseURL, privacy: .public)")
-                        try await transport.connect()
-                        isAuthenticated = true
-                        isConnecting = false
-                        log.info("connect: retry after auto-wake succeeded for \(baseURL, privacy: .public)")
-                        eventStreamClient.startSSE()
-                        return
-                    } catch {
-                        log.error("connect: auto-wake or retry failed for \(baseURL, privacy: .public): \(error)")
-                    }
-                }
-            }
-            #endif
-
-            isConnecting = false
-            httpTransport = nil
-            log.error("connect: transport connection failed for \(baseURL, privacy: .public): \(error)")
-            throw error
-        }
-    }
-
-    // MARK: - Disconnect
-
-    public func disconnect() {
-        disconnectInternal(triggerReconnect: false)
-    }
-
-    func disconnectInternal(triggerReconnect: Bool, cancelAutoWake: Bool = true) {
-        isAuthenticated = false
-
-        #if os(macOS)
-        if cancelAutoWake {
-            autoWakeTask?.cancel()
-            autoWakeTask = nil
-        }
-        #endif
-
-        httpTransport?.disconnect()
-        httpTransport = nil
-        // Stop SSE but preserve subscriber streams — consumers (AppDelegate,
-        // SettingsStore, ConversationRestorer, etc.) start one-shot for-await
-        // loops that must survive reconnects and assistant switches.
-        eventStreamClient.stopSSE()
-
-        isConnected = false
-        isConnecting = false
-        httpPort = nil
-        latestMemoryStatus = nil
-    }
-
-    // MARK: - Auto-Wake
-
-    #if os(macOS)
-    private static let autoWakeCooldown: TimeInterval = 60.0
-
-    private func autoWakeIfDaemonDied() {
-        if isUpdateInProgress {
-            if let expiry = updateExpiresAt, Date() >= expiry {
-                log.warning("auto-wake: planned update expired — clearing stale update state and proceeding")
-                isUpdateInProgress = false
-                updateTargetVersion = nil
-                updateExpiresAt = nil
-                httpTransport?.setUpdateInProgress(false)
-            } else {
-                log.info("auto-wake: skipping — planned service group update in progress")
-                return
-            }
-        }
-
-        guard let wakeHandler,
-              config.transportMetadata.routeMode == .runtimeFlat
-        else { return }
-
-        if let last = lastAutoWakeAttempt,
-           Date().timeIntervalSince(last) < Self.autoWakeCooldown {
-            log.warning("auto-wake: skipping — last attempt was within \(Self.autoWakeCooldown)s cooldown")
-            return
-        }
-
-        lastAutoWakeAttempt = Date()
-
-        autoWakeTask = Task { @MainActor [weak self] in
-            guard let self else { return }
-
-            let reachable = await HealthCheckClient.isReachable(instanceDir: self.config.instanceDir)
-            guard !reachable else {
-                self.lastAutoWakeAttempt = nil
-                return
-            }
-
-            log.info("auto-wake: gateway unreachable — attempting wake")
-            do {
-                try await wakeHandler()
-                guard !Task.isCancelled else {
-                    log.info("auto-wake: cancelled after wake — skipping reconnect")
-                    return
-                }
-                log.info("auto-wake: wake succeeded, reconnecting")
-                try await self.connectImpl(cancelAutoWake: false)
-                guard !Task.isCancelled else {
-                    log.info("auto-wake: cancelled after connect — abandoning")
-                    return
-                }
-                log.info("auto-wake: reconnect succeeded")
-            } catch {
-                log.error("auto-wake: failed: \(error)")
-            }
-            self.autoWakeTask = nil
-        }
-    }
-    #endif
-
-    // MARK: - Token Update
 
     /// Push a new bearer token to the active HTTP transport.
     public func updateTransportBearerToken(_ token: String) {
-        httpTransport?.updateBearerToken(token)
+        connectionManager.updateBearerToken(token)
     }
 
     // MARK: - Version Compatibility
@@ -466,9 +288,6 @@ public final class DaemonStatus: ObservableObject, DaemonStatusProtocol {
     /// Inject a synthetic server message into the event stream. The message
     /// is processed by the state pre-processor and then broadcast to all
     /// subscribers, exactly as if it arrived via SSE.
-    ///
-    /// Used by focused clients (e.g. `AppsClient`) that need to trigger UI
-    /// updates (surface show) without a real SSE event.
     public func injectMessage(_ message: ServerMessage) {
         handleServerMessage(message)
         eventStreamClient.broadcastMessage(message)
@@ -479,7 +298,6 @@ public final class DaemonStatus: ObservableObject, DaemonStatusProtocol {
     /// Handle server messages that update DaemonStatus state. Called synchronously
     /// before the message is broadcast to subscribers.
     private func handleServerMessage(_ message: ServerMessage) {
-        // Handle daemon status updates
         if case .daemonStatus(let status) = message {
             httpPort = status.httpPort.flatMap { Int(exactly: $0) }
             if let version = status.version {
@@ -494,7 +312,7 @@ public final class DaemonStatus: ObservableObject, DaemonStatusProtocol {
                     self.isUpdateInProgress = false
                     self.updateTargetVersion = nil
                     self.updateExpiresAt = nil
-                    self.httpTransport?.setUpdateInProgress(false)
+                    self.connectionManager.httpTransport?.setUpdateInProgress(false)
                 }
             }
             if let newFingerprint = status.keyFingerprint {
@@ -509,36 +327,28 @@ public final class DaemonStatus: ObservableObject, DaemonStatusProtocol {
             }
         }
 
-        // Handle service group update lifecycle
         switch message {
         case .serviceGroupUpdateStarting(let msg):
             self.isUpdateInProgress = true
             self.updateTargetVersion = msg.targetVersion
             self.updateExpiresAt = Date().addingTimeInterval(msg.expectedDowntimeSeconds * 2)
-            self.httpTransport?.setUpdateInProgress(true)
+            self.connectionManager.httpTransport?.setUpdateInProgress(true)
             log.info("Service group update starting — target: \(msg.targetVersion, privacy: .public), expected downtime: \(msg.expectedDowntimeSeconds)s")
         case .serviceGroupUpdateComplete:
             self.isUpdateInProgress = false
             self.updateTargetVersion = nil
             self.updateExpiresAt = nil
-            self.httpTransport?.setUpdateInProgress(false)
+            self.connectionManager.httpTransport?.setUpdateInProgress(false)
         case .modelInfo(let msg):
             currentModel = msg.model
             latestModelInfo = msg
         case .memoryStatus(let msg):
             latestMemoryStatus = msg
         case .authResult(let result):
-            isAuthenticated = result.success
-
+            connectionManager.isAuthenticated = result.success
         default:
             break
         }
-    }
-
-    deinit {
-        #if os(macOS)
-        autoWakeTask?.cancel()
-        #endif
     }
 }
